ISSUE-02: moved shopping_cart.py to package pythonsupermarket. adapted dependent files that use it.
