+++ pythonsupermarket/tests/test_supermarket.py
@@ -1,7 +1,7 @@
 import pytest
 
 from pythonsupermarket.model_objects import Product, SpecialOfferType, ProductUnit
-from shopping_cart import ShoppingCart
+from pythonsupermarket.shopping_cart import ShoppingCart
 from teller import Teller
 from tests.fake_catalog import FakeCatalog
